Small file refactor: move blueprint to own file
