+++ crates/re_viewer/src/ui/mod.rs
@@ -1,36 +1,37 @@
 mod annotations;
-pub use annotations::{Annotations, DefaultColor};
-
 mod auto_layout;
-
-mod viewport;
-pub(crate) use self::viewport::{Blueprint, SpaceViewId};
-
-mod space_view;
-use self::space_view::SpaceView;
-
+mod blueprint;
+mod class_description_ui;
+mod image_ui;
 mod scene;
-use self::scene::SceneQuery;
-
 mod selection_history;
 mod selection_history_ui;
-pub use self::selection_history::{HistoricalSelection, SelectionHistory};
+mod space_view;
+mod view_plot;
+mod view_tensor;
+mod view_text;
+mod viewport;
 
 pub(crate) mod data_ui;
 pub(crate) mod event_log_view;
 pub(crate) mod kb_shortcuts;
+pub(crate) mod memory_panel;
 pub(crate) mod selection_panel;
 pub(crate) mod time_panel;
 
 pub mod view_spatial;
 
-mod class_description_ui;
-mod image_ui;
-mod view_plot;
-mod view_tensor;
-mod view_text;
+// ----
 
-pub(crate) mod memory_panel;
+use self::scene::SceneQuery;
+
+pub(crate) use self::blueprint::Blueprint;
+pub(crate) use self::space_view::{SpaceView, SpaceViewId};
+
+pub use self::annotations::{Annotations, DefaultColor};
+pub use self::selection_history::{HistoricalSelection, SelectionHistory};
+
+// ----
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub(crate) enum Preview {
